updated number data types from Int to Long

+++ src/main/kotlin/data/Language.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Language(
     @SerializedName("lang_id")
-    val langId: Int,
+    val langId: Long,
     @SerializedName("lang_iso")
     val langIso: String,
     @SerializedName("lang_name")
